--- src/data/banner.json
@@ -1,6 +1,0 @@
-{
-    "PromotionBanner1": {
-      "name": "PromotionBanner1",
-      "image": "/assets/banner/PromotionBanner1.svg"
-    }
-  }
