+++ src/pages/Main/assets/ArrowLeft/index.jsx
@@ -1,0 +1,10 @@
+export default function ArrowLeft() {
+    return (
+        <svg width="50" height="50" viewBox="0 0 50 50" fill="none" xmlns="http://www.w3.org/2000/svg">
+<path d="M0 25.0001C0 18.3697 2.63392 12.0109 7.32233 7.32245C12.0107 2.63404 18.3696 0.00012207 25 0.00012207C28.283 0.00012207 31.5339 0.646766 34.5671 1.90313C37.6002 3.1595 40.3562 5.00099 42.6777 7.32245C44.9991 9.64392 46.8406 12.3999 48.097 15.433C49.3534 18.4662 50 21.7171 50 25.0001C50 31.6305 47.3661 37.9894 42.6777 42.6778C37.9893 47.3662 31.6304 50.0001 25 50.0001C21.717 50.0001 18.4661 49.3535 15.4329 48.0971C12.3998 46.8408 9.6438 44.9993 7.32233 42.6778C2.63392 37.9894 0 31.6305 0 25.0001ZM40 22.5001H20L28.75 13.7501L25.2 10.2001L10.4 25.0001L25.2 39.8001L28.75 36.2501L20 27.5001H40V22.5001Z" fill="#3B6EF1"/>
+</svg>
+
+        
+
+    );
+}
