+++ src/pages/Main/assets/ArrowRight/index.jsx
@@ -1,0 +1,10 @@
+export default function ArrowLeft() {
+    return (
+<svg width="50" height="50" viewBox="0 0 50 50" fill="none" xmlns="http://www.w3.org/2000/svg">
+<path d="M50 25.0001C50 18.3697 47.3661 12.0109 42.6777 7.32245C37.9893 2.63404 31.6304 0.00012207 25 0.00012207C21.717 0.00012207 18.4661 0.646766 15.4329 1.90313C12.3998 3.1595 9.6438 5.00099 7.32233 7.32245C5.00087 9.64392 3.15938 12.3999 1.90301 15.433C0.646642 18.4662 0 21.7171 0 25.0001C0 31.6305 2.63392 37.9894 7.32233 42.6778C12.0107 47.3662 18.3696 50.0001 25 50.0001C28.283 50.0001 31.5339 49.3535 34.5671 48.0971C37.6002 46.8408 40.3562 44.9993 42.6777 42.6778C47.3661 37.9894 50 31.6305 50 25.0001ZM10 22.5001H30L21.25 13.7501L24.8 10.2001L39.6 25.0001L24.8 39.8001L21.25 36.2501L30 27.5001H10V22.5001Z" fill="#3B6EF1"/>
+</svg>
+
+        
+
+    );
+}
